Adapted using old sources for qt app, created classes for paint widgets
