+++ BattleshipGame/main.cpp
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "main_window.h"
 #include <QApplication>
 
 int main(int argc, char *argv[])
