+++ BattleshipGame/ship.h
@@ -1,32 +1,28 @@
-#include <iostream>
-#include "battleshipGame.h"
+#ifndef SHIP_H
+#define SHIP_H
+
+#include "square.h"
+#include <vector>
 
 using namespace std;
-using namespace battleshipGameNamespace;
 
-Ship :: Ship(Square startSquare, int shipSize, bool orientation) {
-	if (orientation) {
-		for (int x = startSquare.getX(); x < startSquare.getX() + shipSize; x++) {
-			Square square(x, startSquare.getY());
-			this->squares.push_back(square);
-		}
-	} else {
-		for (int y = startSquare.getY(); y < startSquare.getY() + shipSize; y++) {
-			Square square(startSquare.getX(), y);
-			this->squares.push_back(square);
-		}
-	}
-	this->hP = shipSize;
+namespace battleshipGame {
+
+class Ship {
+
+    private:
+
+    vector <Square> squares;
+    int hP;
+
+    public:
+
+    Ship(Square startSquare, int shipSize, bool orientation);
+    int getHP();
+    void setHP(int);
+    vector <Square> getSquares();
+};
+
 }
 
-int Ship :: getHP() {
-	return this->hP;
-}
-
-void Ship :: setHP(int hP) {
-	this->hP = hP;
-}
-
-vector <Square> Ship :: getSquares() {
-	return this->squares;
-}
+#endif // SHIP_H
