+++ BattleshipGame/shot.h
@@ -1,0 +1,14 @@
+#ifndef SHOT_H
+#define SHOT_H
+
+namespace battleshipGame {
+
+enum class Shot {
+    BESIDE,
+    INJURE,
+    SINK,
+};
+
+}
+
+#endif // SHOT_H
